Addition of 2FA + patch settings

+++ frontend/js/components/userCard/userCard.js
@@ -1,7 +1,10 @@
-import { fetchAuth, fetchUpload, fetchUser, loadHTMLComponent } from "../../api/fetchData.js";
+import { fetchAuth, loadHTMLComponent } from "../../api/fetchData.js";
 import { navigateTo } from "../../router.js";
+import { closeAlertAvatar, closeAlertInfo, setupSettings, closeAlert2FA, clearSettings } from "./utils.js";
+import { disable2FA, enable2FA, updateMenu2FA, checkConfirmationCode } from "./menu2FA.js";
+import { saveAvatar, saveInfo } from "./menuInfo.js";
 import interactiveSocket from '../../pages/home/socket.js'
-import { closeAlertAvatar, closeAlertInfo, setupSettings, removeAllAlerts, noChangeMadeAlert } from "./utils.js";
+
 
 export async function userCardComponent() {
     try {
@@ -17,85 +20,16 @@ export async function userCardListener() {
     document.getElementById('saveInfo').addEventListener('click', saveInfo)
     document.getElementById('saveAvatar').addEventListener('click', saveAvatar)
     document.getElementById('userSettingsModal').addEventListener('show.bs.modal', setupSettings)
-    document.getElementById('userSettingsModal').addEventListener('hide.bs.modal', function (event) {
-        console.log('Settings Modal is about to be hide')
-        document.getElementById('avatarInput').value = ''
-        
-    });
+    document.getElementById('userSettingsModal').addEventListener('hide.bs.modal', clearSettings);
+    document.getElementById('disable2FA').addEventListener('click', disable2FA)
+    document.getElementById('enable2FA').addEventListener('click', enable2FA)
     document.getElementById('btnErrorAvatar').addEventListener('click', closeAlertAvatar)
     document.getElementById('btnErrorInfo').addEventListener('click', closeAlertInfo)
-
+    document.getElementById('closeAlert2FA').addEventListener('click', closeAlert2FA)
+    document.getElementById('2FAForm').addEventListener('submit', checkConfirmationCode)
     settingsListener()
 }
 
-async function displayAlertStatus(response, type) {
-    const alert = document.getElementById('alertError' + type);
-    const alertText = document.getElementById('messageError' + type);
-
-    let message = ""
-    removeAllAlerts(alert);
-    if (response.status == 413) {
-        alert.classList.add('alert-danger');
-        message = "File is too big, max size is 1MB"
-    } else if (response.status == 200) {
-        const dataSuccess = await response.json();
-        document.getElementById('nickname').innerText = dataSuccess.user.nickname;
-        document.getElementById('email').innerText = dataSuccess.user.email;
-        alert.classList.add('alert-success');
-        message = "Your " + type + " has been updated"
-    } else {
-        message = await response.text();
-        alert.classList.add('alert-danger');
-    }
-
-    alert.classList.remove('hide');
-    alert.classList.add('show');
-    alertText.textContent = message
-}
-
-
-async function saveAvatar() {
-    const formData = new FormData();
-    const avatarInput = document.getElementById('avatarInput').files[0];
-    if (avatarInput) {
-        formData.append('avatar', avatarInput);
-    }
-    if (formData.has('avatar')) {
-        const response = await fetchUpload('POST', formData);
-        if (!response) { return }
-        displayAlertStatus(response, 'Avatar')
-    } else {
-        noChangeMadeAlert('messageErrorAvatar', 'alertErrorAvatar');
-    }
-}
-
-
-async function saveInfo() {
-    const objectData = new Object();
-
-    const alert = document.getElementById('alertErrorInfo');
-    removeAllAlerts(alert);
-    closeAlertInfo()
-
-    const nicknameInput = document.getElementById('nicknameInput').value;
-    const userNickname = document.getElementById('nickname').innerText;
-    if (userNickname != nicknameInput) {
-        objectData.nickname = nicknameInput;
-    }
-    const emailInput = document.getElementById('emailInput').value;
-    const userEmail = document.getElementById('email').innerText;
-    if (userEmail != emailInput) {
-        objectData.email = emailInput;
-    }
-
-    if (Object.keys(objectData).length > 0) {
-        const response = await fetchUser('PATCH', null, objectData);
-        if (!response) { return }
-        displayAlertStatus(response, 'Info')
-    } else {
-        noChangeMadeAlert('messageErrorInfo', 'alertErrorInfo');
-    }
-}
 
 function settingsListener() {
     document.querySelectorAll('.left-column-settings i').forEach(function(icon) {
@@ -118,6 +52,7 @@ function settingsListener() {
 export async function logoutUser() {
     console.log('logout!')
     const logoutResponse = await fetchAuth('POST', 'logout/')
+    if (!logoutResponse) { return }
     if (logoutResponse.status == 200) {
         sessionStorage.clear()
         navigateTo('/')
@@ -126,6 +61,7 @@ export async function logoutUser() {
     return ;
 }
 
+
 // Call in home.js
 export async function displayUserCard(meUser) {
     let userContainer = document.getElementById('own-user-card');
@@ -133,6 +69,7 @@ export async function displayUserCard(meUser) {
     let userCard = await userCardComponent();
     userContainer.appendChild(userCard);
     userCardListener(); // enable js on the userCard
+    updateMenu2FA(meUser);
     updateUserCard(meUser);
 }
